% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{comp_indel_mat_d}
\alias{comp_indel_mat_d}
\title{Compute the D matrix for indel calculation.}
\usage{
comp_indel_mat_d(pwm, stat_dist, insertion_len)
}
\arguments{
\item{pwm}{The Position Weight Matrix with dimension motif length x 4.}

\item{stat_dist}{A vector for the stationary distribution for the Markov Chain.}

\item{insertion_len}{An integer for the insertion length.}
}
\value{
A matrix of the same dimension as PWM.
}
\description{
Compute the D matrix for indel calculation.
}
