% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ComputeMotifScore}
\alias{ComputeMotifScore}
\title{Compute the scores for SNP effects on motifs.}
\usage{
ComputeMotifScore(motif.lib, snp.info, ncores = 1)
}
\arguments{
\item{motif.lib}{A list object with the output format of function 'LoadMotifLibrary'.}

\item{snp.info}{A list object with the output format of function 'LoadSNPData'.}

\item{ncores}{An integer for the number of parallel process. Default: 1.}
}
\value{
A data.table containing:
\tabular{ll}{
motif \tab Name of the motif.\cr
snpid \tab The SNP id.\cr
ref_seq \tab The nucleobase sequence for the reference allele.\cr
snp_seq \tab The nucleobase sequence for the SNP allele.\cr
ref_seq_rev \tab The nucleobase sequence for the reference allele on the reverse strand.\cr
snp_seq_rev \tab The nucleobase sequence for the SNP allele on the reverse strand.\cr
motif_len \tab Length of the motif.\cr
ref_start, ref_end, ref_strand \tab Location of the best matching subsequence on the reference allele.\cr
snp_start, snp_end, snp_strand \tab Location of the best matching subsequence on the SNP allele.\cr
log_lik_ref \tab Log-likelihood score for the reference allele.\cr
log_lik_snp \tab Log-likelihood score for the SNP allele.\cr
log_lik_ratio \tab The log-likelihood ratio.\cr
log_enhance_odds \tab Difference in log-likelihood ratio between SNP allele and reference allele based on the best matching subsequence on the reference allele.\cr
log_reduce_odds \tab Difference in log-likelihood ratio between reference allele and SNP allele based on the best matching subsequence on the SNP allele.\cr
ref_match_seq \tab Best matching subsequence on the reference allele.\cr
snp_match_seq \tab Best matching subsequence on the SNP allele.\cr
ref_match_seq \tab Subsequence on the reference allele corresponding to the best matching location on the SNP allele.\cr
snp_match_seq \tab Subsequence on the SNP allele corresponding to the best matching location on the reference allele.\cr
}
Each component is a matrix, with columns representing the motifs and the rows representing the SNPs.
}
\description{
Compute the log-likelihood scores for motifs.
}
\details{
TODO.
}
\examples{
data(example)
ComputeMotifScore(motif_library, snpInfo, ncores = 2)
}
\author{
Chandler Zuo\email{zuo@stat.wisc.edu}
}

