% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ComputePValues}
\alias{ComputePValues}
\title{Compute p values.}
\usage{
ComputePValues(motif.lib, snp.info, motif.scores, ncores)
}
\arguments{
\item{motif.lib}{A list object with the output format of function 'LoadMotifLibrary'.}

\item{snp.info}{A list object with the output format of function 'LoadSNPData'.}

\item{motif.scores}{Log likelihood scores.}

\item{ncores}{An integer for the number of parallel process. Default: 1.}
}
\value{
A data.table with the following columns:
\tabular{ll}{
pval \tab P values.\cr
motif \tab Motif names.\cr
type \tab Type of the test, "a1", "a2", or "diff".}
}
\description{
TODO.
}
\details{
TODO.
}
\examples{
\dontrun{}
}
\author{
Chandler Zuo\email{zuo@stat.wisc.edu}
}

