% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{GetIUPACSequence}
\alias{GetIUPACSequence}
\title{Get the IUPAC sequence of a motif.}
\usage{
GetIUPACSequence(pwm, prob = 0.25)
}
\arguments{
\item{pwm}{The position weight matrix, with the columns representing 
A, C, G, T.}

\item{prob}{The probability threshold. Default: 0.25.}
}
\value{
A character string.
}
\description{
Convert the posotion weight matrix of a motif to the IUPAC 
sequence.
}
\examples{
data(example)
GetIUPACSequence(motif_library[[1]], prob = 0.2)
}
\author{
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}, Chandler Zuo 
\email{chandler.c.zuo@gmail.com}
}
