% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jaspar_motif}
\alias{jaspar_motif}
\title{A motif library containing 593 motifs downloaded from 
http://jaspar.genereg.net/html/DOWNLOAD/JASPAR_CORE/pfm/nonredundant/pfm_all
.txt.}
\format{A list object.}
\description{
This motif library can be loaded by 'data(jaspar_library)'.
}
\author{
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}, Chandler Zuo 
\email{chandler.c.zuo@gmail.com}
}
