% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LoadMotifLibrary}
\alias{LoadMotifLibrary}
\title{Load position weight matrices.}
\usage{
LoadMotifLibrary(filename)
}
\arguments{
\item{filename}{A file containing MEME format: http://meme.nbcr.net/meme/doc/meme-format.html.}
}
\value{
A list object of two components:
\tabular{ll}{
matrix \tab A list of position weight matrices.\cr
prior \tab A vector of the prior distribution for ACGT.\cr}
}
\description{
Load the file for position weight matrices for motifs.
}
\details{
This function reads the MEME formatted file containing motif information and convert them into a list of position weight matrices.
}
\examples{
\dontrun{
pwms <- LoadMotifLibrary("/p/keles/ENCODE-CHARGE/volume1/ENCODE-Motifs/encode_motifs_for_fimo.txt")
}
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

