% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_encode}
\alias{motif_encode}
\title{A motif library provided by the ENCODE consortium.}
\format{A list object.}
\description{
This motif library is obtained from http://compbio.mit.edu/encode-motifs/motifs-toscan.txt.gz. Loaded by 'data(encode_motif)'.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

