% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_info}
\alias{motif_info}
\title{The information for the motif library downloaded from http://compbio.mit.edu/encode-motifs/motifs.txt.}
\format{A character vector.}
\description{
This is a character vector that be loaded by 'data(encode_motif)'. The names of this vector are the same as the names for 'motif_encode'. The entries of this vector are the corresponding motif information parsed from the raw file.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

