% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_library}
\alias{motif_library}
\title{A sample motif library.}
\format{A list object.}
\description{
A list of the position weight matrices corresponding to motifs.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

