% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_match}
\alias{motif_match}
\title{Composit logo plotting input containing motif scores, the matching
subsequences and the augmented matching subsequences on SNP and reference
allele}
\format{A data.frame object.}
\description{
This data.frame object loaded by 'data(example)' contains
information about MYC_disc1 match to rs53576.
}
\author{
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}, Chandler Zuo
\email{chandler.c.zuo@gmail.com}
}
