% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{p_values_for_motif}
\alias{p_values_for_motif}
\title{Compute p-values for a single motif.}
\usage{
p_values_for_motif(
  motif.id,
  par.prior,
  par.transition,
  par.motif.lib,
  par.motif.scores,
  par.testing.mc,
  par.loglik.type,
  par.figdir
)
}
\arguments{
\item{motif.id}{The integer index.}

\item{par.prior}{Prior allele distribution.}

\item{par.transition}{Transition matrix for the allele sequence.}

\item{par.motif.lib}{The motif library.}

\item{par.motif.scores}{Scores in the format of the "motif.scores" field returned
by \code{ComputeMotifScores}.}

\item{par.testing.mc}{See testing.mc in \code{ComputePValues}.}

\item{par.loglik.type}{A string for the log-lik type, "max", "mean" or "median".}

\item{par.figdir}{A string for the path to output figures.}
}
\value{
A list object.
}
\description{
Compute p-values for a single motif.
}
