% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{plotMotifMatch}
\alias{plotMotifMatch}
\title{Plot sequence logos of the position weight matrix of the motif and
sequences of its corresponding best matching augmented subsequence on the
reference and SNP allele.}
\usage{
plotMotifMatch(motif.match, motif.lib, cex.main = 2, ...)
}
\arguments{
\item{motif.match}{a single row ofdtMotifMatch output in data.frame format}

\item{motif.lib}{A list of position weight matrices}

\item{cex.main}{The size of the main title.}

\item{...}{Other parameters passed to plotMotifLogo.}
}
\value{
Sequence logo stacks: Reference subsequences, sequence logo of
reference allele matching potision weight matrix, SNP subsequences, sequence
logo of SNP allele matching potision weight matrix
}
\description{
Plot the best matching augmented subsequences on the reference
and SNP alleles. Plot sequence logos of the position weight matrix of the
motif to the corresponding positions of the best matching subsequences on the
 references and SNP alleles.
}
\examples{
data(example)
plotMotifMatch(motif_match, motif.lib = motif_library)
}
\author{
Sunyoung Shin\email{sunyoung.shin@utdallas.edu}
}
