% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior}
\alias{prior}
\title{Default stationary distribution for nucleotide sequences in the reference genome.}
\format{A numeric vector.}
\description{
This parameter is fitted using 61bp windowns around the SNPs in the NHGRI catalog. Loaded by 'data(default_par)'.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}
