% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pval_with_less_var}
\alias{pval_with_less_var}
\title{Select p-value estimates with less variance.}
\usage{
pval_with_less_var(pval_mat)
}
\arguments{
\item{pval_mat}{A matrix with 4 columns.
Column 1 uses form A
Column 2 is the estimated variance of Column 1
Column 3 uses form B
Column 4 is the estimated variance of Column 3.}
}
\value{
A matrix of 2 columns. Column 1 is the p-value estimate. Column 2 is the estimated variance.
}
\description{
Importance sampling based p-values can be computed in two forms
 A) sum_i weight_i*X_i / N; B) sum_i weight_i*X_i / sum_i weight_i.
 This util function picks the better form.
}
