% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_tbl}
\alias{snp_tbl}
\title{A data frame for SNP information.}
\format{A data.frame object.}
\description{
This data frame is loaded by 'data(example)'. It is a table including the following columns:
\tabular{ll}{
chr \tab The chromosome.\cr
snp \tab The SNP location coordinate.\cr
snpid \tab The SNP label.\cr
a1,a2 \tab The nucleotide on the reference and SNP allele.\cr
}
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}
