% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transition}
\alias{transition}
\title{Default transition probability matrix for nucleotide sequences in the
reference genome.}
\format{
A 4 by 4 numeric matrix.
}
\description{
This parameter is fitted using 61bp windowns around the SNPs in
the NHGRI catalog. Loaded by 'data(default_par)'.
}
\author{
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}, Chandler Zuo
\email{chandler.c.zuo@gmail.com}
}
