% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ComputeMotifScore}
\alias{ComputeMotifScore}
\title{Compute the scores for SNP effects on motifs.}
\usage{
ComputeMotifScore(motif.lib, snp.info, ncores = 1)
}
\arguments{
\item{motif.lib}{A list object with the output format of function 'LoadMotifLibrary'.}

\item{snp.info}{A list object with the output format of function 'LoadSNPData'.}

\item{ncores}{An integer for the number of parallel process. Default: 1.}
}
\value{
A list object with the following components:
\tabular{ll}{
log_lik_ratio \tab The log-likelihood ratio.\cr
on_odds \tab Enhanced log-odds for motif binding by SNP.\cr
off_odds \tab Reduced log-odds for motif binding by SNP.\cr
match_a1 \tab The starting positions of the best match for a1 sequences. Negative numbers correspond to the reverse strand.\cr
match_a2 \tab The starting positions of the best match for a2 sequences. Negative numbers correspond to the reverse strand.\cr
}
Each component is a matrix, with columns representing the motifs and the rows representing the SNPs.
}
\description{
Compute the log-likelihood scores for motifs.
}
\details{
TODO.
}
\examples{
\dontrun{}
}
\author{
Chandler Zuo\email{zuo@stat.wisc.edu}
}

