% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LoadSNPData}
\alias{LoadSNPData}
\title{Load the SNP information and code the genome sequences around the SNP locations.}
\usage{
LoadSNPData(filename, genome.lib = "BSgenome.Hsapiens.UCSC.hg19",
  half.window.size = 30, ...)
}
\arguments{
\item{filename}{A table containing the SNP information. Must contain at least four columns:
\tabular{ll}{
chr \tab chromosome.\cr
snp \tab The nucleotide position of the SNP.\cr
snpid \tab The names of the SNPs.\cr
a1 \tab The deoxyribose at the reference genome.\cr
a2 \tab The deoxyribose at the SNP genome.\cr
}}

\item{genome.lib}{A string of the library name for the genome version. Default: "BSgenome.Hsapiens.UCSC.hg19".}

\item{half.window.size}{An integer for the half window size around the SNP within which the motifs are matched. Default: 30.}
}
\value{
A list object containing the following components:
\tabular{ll}{
sequence_matrix \tab A list of integer vectors representing the deroxyribose sequence around each SNP.\cr
a1 \tab An integer vector for the deroxyribose at the SNP location on the reference genome.\cr
a2 \tab An integer vector for the deroxyribose at the SNP location on the SNP genome.\cr
}
The results are coded as: "A"-1, "C"-2, "G"-3, "T"-4.
}
\description{
Load the SNP data.
}
\details{
TODO.
}
\examples{
\dontrun{LoadSNPData("/p/keles/ENCODE-CHARGE/volume2/SNP/hg19_allinfo.bed")}
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

