% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_encode}
\alias{motif_encode}
\title{A motif library containing 2065 motifs downloaded from http://compbio.mit.edu/encode-motifs/motifs.txt.}
\format{A list object.}
\description{
This motif library can be loaded by 'data(encode_motif)'.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

