% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_library}
\alias{motif_library}
\title{A sample motif library.}
\format{A list object.}
\description{
A list of the position weight matrices corresponding to motifs, loaded by 'data(example)'.
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}
