% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_scores}
\alias{motif_scores}
\title{Scores for the sample snp data computed based on the motif data.}
\format{A list object.}
\description{
This list object contains two fields:
\tabular{ll}{
snp.tbl \tab A data.table containing the sequence of nucleobases around each SNP.\cr
motif.scores \tab A data.table containing the likelihood scores computed for each SNP and each motif.\cr
}
}
\author{
Chandler Zuo zuo@stat.wisc.edu
}

